pyproject.toml
[tool.poetry]
    name        = "bell-avr-vmc-status-module"
    version     = "0.1.0"
    description = "Status LED module for Bell AVR"
    license     = "MIT"
    readme      = "README.md"
    homepage    = "https://github.com/bellflight/AVR-VMC-Status-Module"
    authors     = ["Nathan Vaughn <nvaughn51@gmail.com>"]

[tool.poetry.dependencies]
    python                              = ">=3.9,<3.12"
    adafruit-circuitpython-neopixel-spi = "1.0.7"
    loguru                              = "0.7.0"
    bell-avr-libraries                  = "0.2.0a4"
    # https://github.com/NVIDIA/jetson-gpio/issues/79
    jetson-gpio = { version = "!=2.1.0", markers = "sys_platform == 'linux'" }

[tool.poetry.group.dev.dependencies]
    pre-commit  = "3.3.1"
    pytest-cov  = "4.1.0"
    pytest      = "7.4.0"
    pytest-mock = "3.11.1"

[tool.pyright]
    typeCheckingMode           = "basic"
    venvPath                   = "."
    venv                       = ".venv"
    reportMissingParameterType = true

[tool.ruff]
    ignore = ["E501"]
